+++ CHANGES.rst
@@ -28,6 +28,8 @@ Release history
   from a spatial shape and number of channels. (`#1648`_)
 - Added the ``ConvolutionTranspose`` transform to perform transpose convolution; this
   can be used to achieve various forms of upsampling in deep networks. (`#1648`_)
+- Added ``Conv`` and ``ConvTranspose`` aliases for ``Convolution`` and
+  ``ConvolutionTranspose``. (`#1648`_)
 
 **Fixed**